pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "espidf-mcp"
version = "0.2.0"
description = "MCP server for ESP-IDF development workflow with stage-based management"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ESP-IDF MCP Contributors"}
]
keywords = ["esp-idf", "mcp", "model-context-protocol", "esp32", "embedded"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Hardware",
]

dependencies = [
    "fastmcp>=2.14.2",
    "mcp>=1.10.0",
    "pyserial>=3.5",
    "numpy>=1.24.0",  # For accurate percentile calculation in metrics
    "filelock>=3.13.0",  # For cross-platform file locking
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "mypy>=1.10.0",
]

[project.scripts]
espidf-mcp = "cli:main"
espidf-mcp-python = "__main__:main"

[project.urls]
Homepage = "https://github.com/your-org/espidf-mcp"
Repository = "https://github.com/your-org/espidf-mcp"
Issues = "https://github.com/your-org/espidf-mcp/issues"

[tool.setuptools]
py-modules = [
    "cli",
    "project",
    "server",
    "workflow_server",
    "__init__",
    "__main__",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_tools*", "tools*", "checkers*", "workflow*", "config*", "tests*", "observability*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
show_error_codes = true
pretty = true
